Renames *connect_step* to *sequence_step* and moves this family to the framed module. These should be able to be used in both ironrdp-async/src/connector.rs and ironrdp-acceptor/src/lib.rs (however the latter has not been manually tested as of this commit).
