+++ crates/ironrdp-acceptor/src/lib.rs
@@ -1,8 +1,8 @@
 #[macro_use]
 extern crate tracing;
 
-use ironrdp_async::{Framed, FramedRead, FramedWrite, StreamWrapper};
-use ironrdp_connector::{custom_err, ConnectorResult, Sequence, Written};
+use ironrdp_async::{single_sequence_step, Framed, FramedRead, FramedWrite, StreamWrapper};
+use ironrdp_connector::ConnectorResult;
 use ironrdp_pdu::write_buf::WriteBuf;
 
 mod channel_connection;
@@ -41,7 +41,7 @@ where
             return Ok(result);
         }
 
-        single_accept_state(&mut framed, acceptor, &mut buf).await?;
+        single_sequence_step(&mut framed, acceptor, &mut buf).await?;
     }
 }
 
@@ -59,48 +59,6 @@ where
             return Ok((framed, result));
         }
 
-        single_accept_state(&mut framed, acceptor, &mut buf).await?;
+        single_sequence_step(&mut framed, acceptor, &mut buf).await?;
     }
-}
-
-async fn single_accept_state<S>(
-    framed: &mut Framed<S>,
-    acceptor: &mut Acceptor,
-    buf: &mut WriteBuf,
-) -> ConnectorResult<Written>
-where
-    S: FramedRead + FramedWrite,
-{
-    buf.clear();
-
-    let written = if let Some(next_pdu_hint) = acceptor.next_pdu_hint() {
-        debug!(
-            acceptor.state = acceptor.state().name(),
-            hint = ?next_pdu_hint,
-            "Wait for PDU"
-        );
-
-        let pdu = framed
-            .read_by_hint(next_pdu_hint)
-            .await
-            .map_err(|e| custom_err!("read frame by hint", e))?;
-
-        trace!(length = pdu.len(), "PDU received");
-
-        acceptor.step(&pdu, buf)?
-    } else {
-        acceptor.step_no_input(buf)?
-    };
-
-    if let Some(response_len) = written.size() {
-        debug_assert_eq!(buf.filled_len(), response_len);
-        let response = buf.filled();
-        trace!(response_len, "Send response");
-        framed
-            .write_all(response)
-            .await
-            .map_err(|e| custom_err!("write all", e))?;
-    }
-
-    Ok(written)
 }
